# Conflicts:
#	.gitignore
#	pom.xml
#	src/main/java/com/watermelon/config/DruidConfig.java
#	src/main/java/com/watermelon/config/ShiroConfig.java
#	src/main/java/com/watermelon/controller/AdminController.java
#	src/main/java/com/watermelon/controller/CourseManageController.java
#	src/main/java/com/watermelon/controller/IndexController.java
#	src/main/java/com/watermelon/controller/UserController.java
#	src/main/java/com/watermelon/entity/Teacher.java
#	src/main/java/com/watermelon/service/PermissionServiceImpl.java
#	src/main/java/com/watermelon/service/RoleServiceImpl.java
#	src/main/java/com/watermelon/service/UserServiceImpl.java
#	src/main/resources/application.yml
#	src/main/resources/mapper/permissionMapper.xml
#	src/main/resources/mapper/roleMapper.xml
#	src/main/resources/mapper/userMapper.xml

+++ src/main/java/com/watermelon/service/PermissionServiceImpl.java
@@ -30,16 +30,25 @@ public class PermissionServiceImpl implements PermissionService {
 
     @Override
     public void addPermission(Permission permission) {
-        permissionMapper.addPermission(permission);
+        Permission permission1 = permissionMapper.getPermissionById(permission.getId());
+        if (permission1==null) {
+            permissionMapper.addPermission(permission);
+        }
     }
 
     @Override
     public void updatePermission(Permission permission) {
-        permissionMapper.updatePermission(permission);
+        Permission permission1 = permissionMapper.getPermissionById(permission.getId());
+        if (permission1!=null) {
+            permissionMapper.updatePermission(permission);
+        }
     }
 
     @Override
     public void deletePermission(int id) {
-        permissionMapper.deletePermission(id);
+        Permission permission1 = permissionMapper.getPermissionById(id);
+        if (permission1!=null) {
+            permissionMapper.deletePermission(id);
+        }
     }
 }
